refactor: Move MockNodeKeyManager back to production code
