+++ packages/dds/tree/src/feature-libraries/node-key/index.ts
@@ -3,13 +3,12 @@
  * Licensed under the MIT License.
  */
 
+export { MockNodeKeyManager } from "./mockNodeKeyManager.js";
 export {
 	compareLocalNodeKeys,
 	LocalNodeKey,
 	StableNodeKey,
 	nodeKeyTreeIdentifier,
 } from "./nodeKey.js";
-
 export { NodeKeyIndex } from "./nodeKeyIndex.js";
-
 export { createNodeKeyManager, isStableNodeKey, NodeKeyManager } from "./nodeKeyManager.js";
