+++ packages/dds/tree/src/feature-libraries/node-key/mockNodeKeyManager.ts
@@ -6,13 +6,9 @@
 import { assert } from "@fluidframework/core-utils/internal";
 import type { SessionSpaceCompressedId, StableId } from "@fluidframework/id-compressor";
 import { assertIsStableId } from "@fluidframework/id-compressor/internal";
-import {
-	isStableNodeKey,
-	LocalNodeKey,
-	NodeKeyManager,
-	StableNodeKey,
-} from "../feature-libraries/index.js";
-import { brand, extractFromOpaque, fail } from "../util/index.js";
+import { LocalNodeKey, StableNodeKey } from "./nodeKey.js";
+import { isStableNodeKey, NodeKeyManager } from "./nodeKeyManager.js";
+import { brand, extractFromOpaque, fail } from "../../util/index.js";
 
 /**
  * Mock {@link NodeKeyManager} that generates deterministic {@link StableNodeKey}s and {@link LocalNodeKey}s.
