+++ packages/dds/tree/src/test/feature-libraries/flex-tree/utils.ts
@@ -18,10 +18,10 @@ import {
 	FlexAllowedTypes,
 	FlexFieldKind,
 	FlexTreeSchema,
+	MockNodeKeyManager,
 } from "../../../feature-libraries/index.js";
 import { TreeContent } from "../../../shared-tree/index.js";
 import { MockTreeCheckout, forestWithContent } from "../../utils.js";
-import { MockNodeKeyManager } from "../../mockNodeKeyManager.js";
 
 export function getReadonlyContext(forest: IEditableForest, schema: FlexTreeSchema): Context {
 	return getTreeContext(schema, new MockTreeCheckout(forest), new MockNodeKeyManager());
