+++ packages/dds/tree/src/test/feature-libraries/node-key/nodeKey.spec.ts
@@ -14,10 +14,10 @@ import {
 	StableNodeKey,
 	compareLocalNodeKeys,
 	createNodeKeyManager,
+	MockNodeKeyManager,
 } from "../../../feature-libraries/index.js";
 import { ISharedTree } from "../../../shared-tree/index.js";
 import { TestTreeProvider } from "../../utils.js";
-import { MockNodeKeyManager } from "../../mockNodeKeyManager.js";
 
 /**
  * Acquire an {@link IIdCompressor} via unsavory means.
