+++ packages/dds/tree/src/test/shared-tree/schematizingTreeView.spec.ts
@@ -10,8 +10,9 @@ import { UsageError } from "@fluidframework/telemetry-utils/internal";
 import {
 	FieldKinds,
 	FlexFieldSchema,
-	SchemaBuilderBase,
 	intoStoredSchema,
+	MockNodeKeyManager,
+	SchemaBuilderBase,
 } from "../../feature-libraries/index.js";
 // eslint-disable-next-line import/no-internal-modules
 import { UpdateType } from "../../shared-tree/schematizeTree.js";
@@ -25,7 +26,6 @@ import { SchemaFactory, TreeConfiguration } from "../../simple-tree/index.js";
 import { toFlexConfig, toFlexSchema } from "../../simple-tree/toFlexSchema.js";
 import { disposeSymbol } from "../../util/index.js";
 import { checkoutWithContent, createTestUndoRedoStacks, insert } from "../utils.js";
-import { MockNodeKeyManager } from "../mockNodeKeyManager.js";
 
 const schema = new SchemaFactory("com.example");
 const config = new TreeConfiguration(schema.number, () => 5);
