+++ packages/dds/tree/src/test/simple-tree/utils.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-import { NodeKeyManager } from "../../feature-libraries/index.js";
+import { MockNodeKeyManager, NodeKeyManager } from "../../feature-libraries/index.js";
 import {
 	ImplicitFieldSchema,
 	InsertableTreeFieldFromImplicitField,
@@ -13,7 +13,6 @@ import {
 } from "../../simple-tree/index.js";
 // eslint-disable-next-line import/no-internal-modules
 import { getProxyForField } from "../../simple-tree/proxies.js";
-import { MockNodeKeyManager } from "../mockNodeKeyManager.js";
 import { flexTreeWithContent } from "../utils.js";
 
 /**
